+++ infra/url/core/src/test/resources/config/to-be-replaced-fixture.yaml
